enhanced relationship between entities

+++ src/main/java/org/example/parkingmanagementbackend/repository/FloorRepository.java
@@ -3,5 +3,5 @@ package org.example.parkingmanagementbackend.repository;
 import org.example.parkingmanagementbackend.model.Floor;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FloorRepository extends JpaRepository<Floor, Long> {
+public interface FloorRepository extends JpaRepository<Floor, Integer> {
 }
